requirements.txt
black
pyspark
dataclasses
pandas
more_itertools
aiohttp
PyYAML
Pillow
tweepy
ipykernel
openai
tweepy[async]
playwright
jupyter
notebook
itables
pyarrow
fastparquet
matplotlib
networkx
scipy